add delete message rest api endpoint

+++ restApi/src/domain/message/commands/deleteMessageCommandImpl.ts
@@ -6,12 +6,12 @@ export class DeleteMessageCommandImpl implements DeleteMessageCommand {
   public constructor(private readonly messageRepository: MessageRepository) {}
 
   public async deleteMessage(input: DeleteMessageCommandPayload): Promise<void> {
-    const { id } = DeleteMessageCommandPayload.create(input);
+    const { id, userId } = DeleteMessageCommandPayload.create(input);
 
-    console.log('Deleting message...', { id });
+    console.log('Deleting message...', { id, userId });
 
-    await this.messageRepository.deleteOne({ id });
+    await this.messageRepository.deleteOne({ id, user_id: userId });
 
-    console.log('Message deleted.', { id });
+    console.log('Message deleted.', { id, userId });
   }
 }
